+++ apps/entity-service/src/entities/dto/task-update.dto.ts
@@ -1,7 +1,8 @@
-import { IsIn, IsOptional } from "class-validator";
+import { IsIn, IsOptional, IsArray, IsString } from "class-validator";
 import { TASK_STATUSES, TaskStatus } from "@terramatch-microservices/database/constants/status";
 import { ApiProperty } from "@nestjs/swagger";
 import { JsonApiBodyDto, JsonApiDataDto } from "@terramatch-microservices/common/util/json-api-update-dto";
+import { Type } from "class-transformer";
 
 class TaskUpdateAttributes {
   @IsOptional()
@@ -12,8 +13,73 @@ class TaskUpdateAttributes {
     enum: TASK_STATUSES
   })
   status?: TaskStatus | null;
+
+  @IsOptional()
+  @IsString()
+  @ApiProperty({ description: "Specific feedback for the PD", nullable: true, type: String })
+  feedback?: string | null;
+
+  @IsOptional()
+  @IsArray()
+  @Type(() => String)
+  @ApiProperty({
+    description: "UUIDs of site reports to mark as 'Nothing to report'",
+    isArray: true,
+    type: String,
+    nullable: true
+  })
+  siteReportNothingToReportUuid?: string[];
+
+  @IsOptional()
+  @IsArray()
+  @Type(() => String)
+  @ApiProperty({
+    description: "UUIDs of nursery reports to mark as 'Nothing to report'",
+    isArray: true,
+    type: String,
+    nullable: true
+  })
+  nurseryReportNothingToReportUuid?: string[];
 }
 
 export class TaskUpdateBody extends JsonApiBodyDto(
   class TaskData extends JsonApiDataDto({ type: "tasks" }, TaskUpdateAttributes) {}
 ) {}
+
+export class BulkTaskUpdateBody {
+  @IsOptional()
+  @IsIn(TASK_STATUSES)
+  @ApiProperty({
+    description: "Request to change to the status of the given tasks",
+    nullable: true,
+    enum: TASK_STATUSES
+  })
+  status?: TaskStatus | null;
+
+  @IsOptional()
+  @IsString()
+  @ApiProperty({ description: "Specific feedback for the PD", nullable: true, type: String })
+  feedback?: string | null;
+
+  @IsOptional()
+  @IsArray()
+  @Type(() => String)
+  @ApiProperty({
+    description: "UUIDs of site reports to mark as 'Nothing to report'",
+    isArray: true,
+    type: String,
+    nullable: true
+  })
+  siteReportNothingToReportUuid?: string[];
+
+  @IsOptional()
+  @IsArray()
+  @Type(() => String)
+  @ApiProperty({
+    description: "UUIDs of nursery reports to mark as 'Nothing to report'",
+    isArray: true,
+    type: String,
+    nullable: true
+  })
+  nurseryReportNothingToReportUuid?: string[];
+}
