+++ apps/entity-service/src/entities/processors/entity.processor.spec.ts
@@ -132,74 +132,4 @@ describe("EntityProcessor", () => {
       });
     });
   });
-
-  describe("updateBulkApprovalReports", () => {
-    it("should handle site report nothing to report status", async () => {
-      const siteReport = await SiteReportFactory.create();
-      const update = { siteReportNothingToReportUuid: [siteReport.uuid], feedback: "Test feedback" };
-      await ProjectFactory.create();
-      await createProcessor().updateBulkApprovalReports(update, APPROVED);
-      const updatedSiteReport = await SiteReport.findOne({ where: { uuid: siteReport.uuid } });
-      expect(updatedSiteReport?.status).toBe(APPROVED);
-      const auditStatus = await AuditStatus.findOne({
-        where: { auditableType: laravelType(siteReport), auditableId: siteReport.id }
-      });
-      expect(auditStatus).toMatchObject({ status: APPROVED, comment: "Test feedback" });
-    });
-
-    it("should handle nursery report nothing to report status", async () => {
-      const nurseryReport = await NurseryReportFactory.create();
-      const update = { nurseryReportNothingToReportUuid: [nurseryReport.uuid], feedback: "Test feedback" };
-      await ProjectFactory.create();
-      await createProcessor().updateBulkApprovalReports(update, APPROVED);
-      const updatedNurseryReport = await NurseryReport.findOne({ where: { uuid: nurseryReport.uuid } });
-      expect(updatedNurseryReport?.status).toBe(APPROVED);
-      const auditStatus = await AuditStatus.findOne({
-        where: { auditableType: laravelType(nurseryReport), auditableId: nurseryReport.id }
-      });
-      expect(auditStatus).toMatchObject({ status: APPROVED, comment: "Test feedback" });
-    });
-
-    it("should handle both site and nursery reports in a single update", async () => {
-      const siteReport = await SiteReportFactory.create();
-      const nurseryReport = await NurseryReportFactory.create();
-      const update = {
-        siteReportNothingToReportUuid: [siteReport.uuid],
-        nurseryReportNothingToReportUuid: [nurseryReport.uuid],
-        feedback: "Test feedback"
-      };
-      await ProjectFactory.create();
-      await createProcessor().updateBulkApprovalReports(update, APPROVED);
-
-      const [updatedSiteReport, updatedNurseryReport] = await Promise.all([
-        SiteReport.findOne({ where: { uuid: siteReport.uuid } }),
-        NurseryReport.findOne({ where: { uuid: nurseryReport.uuid } })
-      ]);
-
-      expect(updatedSiteReport?.status).toBe(APPROVED);
-      expect(updatedNurseryReport?.status).toBe(APPROVED);
-
-      const [siteAuditStatus, nurseryAuditStatus] = await Promise.all([
-        AuditStatus.findOne({
-          where: { auditableType: laravelType(siteReport), auditableId: siteReport.id }
-        }),
-        AuditStatus.findOne({
-          where: { auditableType: laravelType(nurseryReport), auditableId: nurseryReport.id }
-        })
-      ]);
-
-      expect(siteAuditStatus).toMatchObject({ status: APPROVED, comment: "Test feedback" });
-      expect(nurseryAuditStatus).toMatchObject({ status: APPROVED, comment: "Test feedback" });
-    });
-
-    it("should handle empty report lists", async () => {
-      const update = {
-        siteReportNothingToReportUuid: [],
-        nurseryReportNothingToReportUuid: [],
-        feedback: "Test feedback"
-      };
-      await ProjectFactory.create();
-      await expect(createProcessor().updateBulkApprovalReports(update, APPROVED)).resolves.not.toThrow();
-    });
-  });
 });
